+++ packages/core/CHANGELOG.md
@@ -10,6 +10,7 @@ Versioning].
 
 - Fix mangling IDs followed by child, sibling, and adjacent combinators in CSS.
 - Fix accidental mangling of some non-id attributes in HTML.
+- Support mangling IDs in CSS attribute selectors using `href`.
 
 ## [0.1.9] - 2021-02-12